requirements.txt
python-telegram-bot>=13.7
requests>=2.28.1
python-dotenv>=0.19.2
aiohttp>=3.8.3
